Prediction wrappers for h2o

Prediction wrappers for fitted models with h2o engine that include data conversion, h2o server cleanup, and so on.

Usage

h2o_predict(object, new_data, ...)

h2o_predict_classification(object, new_data, type = "class", ...)

h2o_predict_regression(object, new_data, type = "numeric", ...)

# S3 method for `_H2OAutoML`
predict(object, new_data, id = NULL, ...)

Arguments

object

An object of class model_fit

new_data

A rectangular data object, such as a data frame.

...

Other options passed to h2o::h2o.predict()

type

A single character value or NULL. Possible values are "numeric", "class", "prob", "conf_int", "pred_int", "quantile", "time", "hazard", "survival", or "raw". When NULL, predict() will choose an appropriate value based on the model's mode.

id

Model id in AutoML results.

Value

For type != "raw", a prediction data frame with the same number of rows as new_data. For type == "raw", return original h2o::h2o.predict()

output

Details

For AutoML, prediction is based on the best performing model.

Examples

if (h2o_running()) {
  spec <-
    rand_forest(mtry = 3, trees = 100) %>%
    set_engine("h2o") %>%
    set_mode("regression")

  set.seed(1)
  mod <- fit(spec, mpg ~ ., data = mtcars)
  h2o_predict_regression(mod$fit, new_data = head(mtcars), type = "numeric")

  # using parsnip
  predict(mod, new_data = head(mtcars))
}
#> # A tibble: 6 × 1
#>   .pred
#>   <dbl>
#> 1  20.6
#> 2  20.7
#> 3  22.7
#> 4  20.2
#> 5  17.8
#> 6  19.1